change engine

+++ torch_utils/engine.py
@@ -45,7 +45,7 @@ def train_one_epoch(
     for images, targets in metric_logger.log_every(data_loader, print_freq, header):
         step_counter += 1
         images = list(image.to(device) for image in images)
-        targets = [{k: v.to(device).to(torch.int64) for k, v in t.items()} for t in targets]
+        targets = [{k: v.to(device) for k, v in t.items()} for t in targets]
 
 
         with torch.cuda.amp.autocast(enabled=scaler is not None):
